moved listenToFollowRequests to FollowProvider
