+++ code/app/src/main/java/com/kernelcrew/moodapp/data/FollowProvider.java
@@ -5,14 +5,10 @@ import android.util.Log;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
-import com.google.android.gms.tasks.Tasks;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.kernelcrew.moodapp.utils.NotificationHelper;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -22,6 +18,25 @@ public class FollowProvider {
     private final FirebaseFirestore db = FirebaseFirestore.getInstance();
 
     private FollowProvider() {}
+
+    /**
+     * Listens to changes in the follow requests collection for a specific user.
+     * This method adds a snapshot listener to the follow requests collection and triggers
+     * the provided listener whenever there is a change in the collection (e.g., a new follow
+     * request is added or removed).
+     *
+     * @param userId The unique ID of the user whose follow requests are being monitored.
+     * @param listener The listener that will be triggered whenever there is a change
+     *                 in the follow requests collection.
+     */
+    public static void listenToFollowRequests(String userId, EventListener<QuerySnapshot> listener) {
+        // Set up a Firestore snapshot listener to listen to changes in follow requests
+        FirebaseFirestore.getInstance()
+                .collection("users")
+                .document(userId)
+                .collection("followRequests")
+                .addSnapshotListener(listener); // Returns a listener for updates
+    }
 
     public static FollowProvider getInstance() {
         if (instance == null) {
